+++ signal-sweep/src/signal_sweep/core/handlers/base_handler.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import List
 
-from ..signal_stream import StreamData
+from ..models import StreamData
 
 
 class Handler(ABC):
