+++ signal-sweep/src/signal_sweep/core/handlers/text_handler.py
@@ -10,7 +10,7 @@ import re
 import httpx
 
 from .base_handler import Handler
-from ..signal_stream import StreamData
+from ..models import StreamData
 from ..models import Source
 
 IP_V4_REGEX = r"\b(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\b"
