+++ signal-sweep/src/signal_sweep/core/models.py
@@ -7,3 +7,9 @@ from ..shared.constants import SourceType
 class Source:
     url: str
     type: SourceType
+
+
+@dataclass(frozen=True)
+class StreamData:
+    ip: str
+    source_url: str
