+++ signal-sweep/tests/conftest.py
@@ -14,7 +14,8 @@ import pytest
 import httpx
 import redis.asyncio as redis
 
-from signal_sweep.core.signal_stream import StreamData, SignalStream
+from signal_sweep.core.signal_stream import SignalStream
+from signal_sweep.core.models import StreamData
 from signal_sweep.core.models import Source
 from signal_sweep.shared.constants import SourceType
 from signal_sweep.shared.utils import AsyncProcessPoolExecutor
